requirements.txt
# Core Django & REST
django>=5.1,<5.2
djangorestframework>=3.14
django-filter>=24.2
pillow>=10.0

# Environment & auth
python-decouple>=3.8
djangorestframework-simplejwt>=5.3

# Docs (optional but useful for API testing)
drf-yasg>=1.21

# Agar Celery/periodic tasks use kar rahe ho to yeh add karo (conflict fix ke liye)
# django-celery-beat>=2.6.0   # Uncomment kar do agar zarurat ho, latest versions Django 5.1 support karte hain

# Optional extras (common e-commerce ke liye)
django-cors-headers>=4.0        # Flutter frontend se CORS allow karne ke liye
gunicorn>=22.0                  # Production server (optional)